# frontend/mrsexam-ui/root/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "mrsexam-backend",
    "version": "1.0"
  },
  "paths": {
...
    "/api/CentreCounts/region/{regionNo}": {
      "get": {
        "tags": [
          "CentreCounts"
        ],
        "parameters": [
          {
            "name": "regionNo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/FormBalbodhini/region/{regionNo}": {
      "get": {
        "tags": [
          "FormBalbodhini"
        ],
        "parameters": [
          {
            "name": "regionNo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/FormPrabodh/region/{regionNo}": {
      "get": {
        "tags": [
          "FormPrabodh"
        ],
        "parameters": [
          {
            "name": "regionNo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/FormPrathamik/region/{regionNo}": {
      "get": {
        "tags": [
          "FormPrathamik"
        ],
        "parameters": [
          {
            "name": "regionNo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/FormPraveen/region/{regionNo}": {
      "get": {
        "tags": [
          "FormPraveen"
        ],
        "parameters": [
          {
            "name": "regionNo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/FormPraveshika/region/{regionNo}": {
      "get": {
        "tags": [
          "FormPraveshika"
        ],
        "parameters": [
          {
            "name": "regionNo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/FormSubodh/region/{regionNo}": {
      "get": {
        "tags": [
          "FormSubodh"
        ],
        "parameters": [
          {
            "name": "regionNo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/Regions": {
      "get": {
        "tags": [
          "Regions"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Regions"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Region"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Region"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Region"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Regions/paged": {
      "get": {
        "tags": [
          "Regions"
        ],
        "parameters": [
          {
            "name": "pageNumber",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Regions/{id}": {
      "get": {
        "tags": [
          "Regions"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Regions"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Region"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Region"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Region"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Regions"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Results": {
      "get": {
        "tags": [
          "Results"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Results"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Result"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Result"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Result"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Results/{id}": {
      "get": {
        "tags": [
          "Results"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Results"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Result"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Result"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Result"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Results"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Region": {
        "required": [
          "regionNo"
        ],
        "type": "object",
        "properties": {
          "regionNo": {
            "type": "integer",
            "format": "int32"
          },
          "regionName": {
            "maxLength": 50,
            "type": "string",
            "nullable": true
          },
          "seatRegionNo": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "regionNameEng": {
            "maxLength": 50,
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Result": {
        "type": "object",
        "properties": {
          "resultId": {
            "type": "integer",
            "format": "int32"
          },
          "resultValue": {
            "maxLength": 100,
            "type": "string",
            "nullable": true
          },
          "description": {
            "maxLength": 100,
            "type": "string",
            "nullable": true
          },
          "exam": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}